third_party/WebKit/LayoutTests/http/tests/htmlimports/encoding.html

<!DOCTYPE html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <link rel="import" href="resources/no-encoding.cgi">
    <link rel="import" href="resources/http-encoding.cgi">
</head>
<body>
<script>
description("Test that UTF-8 decoding is applied appropriately to imports");
jsTestIsAsync = true;
window.onload = function() {
    links = document.querySelectorAll('link[rel=import]');
    debug("Default encoding should be UTF-8...");
    shouldBeEqualToString("links[0].import.characterSet", "UTF-8");
    debug("And Content-Type headers should be ignored...");
    shouldBeEqualToString("links[1].import.characterSet", "UTF-8");
    finishJSTest();
};
</script>
</body>
