third_party/WebKit/LayoutTests/http/tests/htmlimports/encoding-expected.txt

Test that UTF-8 decoding is applied appropriately to imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default encoding should be UTF-8...
PASS links[0].import.characterSet is "UTF-8"
And Content-Type headers should be ignored...
PASS links[1].import.characterSet is "UTF-8"
PASS successfullyParsed is true

TEST COMPLETE
